pyproject.toml
    [project]
    name = "python_course"
    version = "0.1.0"
dependencies = [
    "flask>=3.1.2",
    "psycopg2-binary>=2.9.11",
]
